+++ src/ApiCalls/index.js
@@ -1,0 +1,3 @@
+import { getData, getCategoryData } from "./getMethodRequest";
+
+export { getData, getCategoryData}
