+++ src/Context/data.context.js
@@ -1,5 +1,6 @@
-import axios from "axios";
 import {createContext, useState, useContext, useEffect} from "react";
+import { getData, getCategoryData } from "../ApiCalls";
+
 const DataContext =  createContext();
 
 const DataProvider = ({children}) => {
@@ -11,26 +12,7 @@ const DataProvider = ({children}) => {
         getCategoryData(setCategoryList);
     }, []);
 
-    const getCategoryData = async(setCategoryList) => {
-        try{
-            const response = await axios.get("/api/categories");
-            if(response?.data?.categories){
-                setCategoryList(response?.data?.categories)
-            }
-        }catch(error) {
-            console.error(error)
-        }
-    }
-    const getData = async(setVideosList) => {
-        try{
-            const response = await axios.get("/api/videos");
-            if(response?.data?.videos) {
-                setVideosList(response?.data?.videos)
-            }
-        } catch(error){
-            console.error(error);
-        }
-    }
+   
     return (
         <DataContext.Provider value={{videosList, categoryList}}>
             {children}
