+++ src/Context/index.js
@@ -1,0 +1,4 @@
+import { useData } from "./data.context";
+import { useFilter } from "./filter.context";
+
+export {useData, useFilter};
