+++ src/Helper/index.js
@@ -1,0 +1,3 @@
+import { getFilteredBySearchData, getFilteredData } from "./filterFunciton";
+
+export {getFilteredBySearchData, getFilteredData}
